requirements.txt
asttokens==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
executing==2.1.0
idna==3.10
jedi==0.19.2
keyboard==0.13.5
MouseInfo==0.1.3
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pure_eval==0.2.3
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyTelegramBotAPI==4.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pywin32==308
requests==2.32.3
setuptools==75.6.0
six==1.16.0
stack-data==0.6.3
urllib3==2.2.3
wcwidth==0.2.13
